requirements.txt
blinker==1.4
docutils==0.12
feedgenerator==1.8
ghp-import==0.4.1
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.3
pelican==3.6.3
pycrypto==2.6.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
PyYAML==3.12
six==1.10.0
sjcl==0.1.5
Unidecode==0.4.19
